feat: punishments endpoint & organize routes
