+++ src/server/servers/routes/log.ts
@@ -1,0 +1,37 @@
+import { Auxdibot } from '@/interfaces/Auxdibot';
+import checkAuthenticated from '@/server/checkAuthenticated';
+import { Router } from 'express';
+/*
+   Log
+   View the latest log on your server & the log channel.
+*/
+const log = (auxdibot: Auxdibot, router: Router) => {
+   router.get(
+      '/:serverID/log',
+      (req, res, next) => checkAuthenticated(req, res, next),
+      (req, res) => {
+         if (!req.user?.guilds) return res.status(400).json({ error: 'no servers' });
+         const serverID = req.params.serverID;
+         const guildData = req.user.guilds.find((i) => i.id == serverID);
+         const guild = auxdibot.guilds.cache.get(serverID);
+         if (!guildData || !guild) return res.status(404).json({ error: "couldn't find that server" });
+         if (!guildData.owner && !(guildData.permissions & 0x8))
+            return res.status(403).json({ error: 'you are not authorized to edit that server' });
+
+         return auxdibot.database.servers
+            .findFirst({
+               where: { serverID: serverID },
+               select: { serverID: true, latest_log: true, log_channel: true },
+            })
+            .then(async (data) =>
+               data ? res.json({ ...guildData, data }) : res.status(404).json({ error: "couldn't find that server" }),
+            )
+            .catch((x) => {
+               console.error(x);
+               return res.status(500).json({ error: 'an error occurred' });
+            });
+      },
+   );
+   return router;
+};
+export default log;
